catch-up: Mass commit - been too long - Massive I/O completion API improvement! - Per-CPU heap! - Update README - Add spectre mitigation to bootloader - Add resolution to boot menu and do linear framebuffer setup for BIOS/EFI - Handle partial success in stage1 int 13 ah=42h - Add __cxa_atexit and __dso_handle suddenly needed in bootloader - Add GRAPHICS_OUTPUT_PROTOCOL support to EFI declarations - Drastic improvements in far pointer wrapper class - Remove debug spam in bootloader - Add some no-instrumentation versions of some functions - Show CPU usage in debug spam - Massive call trace improvements throughout - Add (some) meltdown bug handling - Calculate ISR entry points instead of using big lookup table - Add FLUSH_TRACE software interrupt - Simplify ISR entry points - Add some support for AVX/AVX-512/etc register fetch for gdbstub, etc - Move physical mapping to 512GB at highest 512GB boundary >= 512GB   before .text - Use smaller instructions for CPU-local data access - Fix misuse of cfi_restore - Pin all threads to a specific CPU - Add ES1370 sound card driver bitfield declarations - Add final to several classes - I/O completions now have storage for a size and error code - Cache PCI config up front and use cache when probing for devices - Report bytes transferred correctly when reading USB pipe - Annotate several PCI drivers with detection parameters on 1st line - Start VT100 terminal emulator - ASAN fixes - Hang in assertion failure for debugging convenience - Callout when heap is ready - Add allocator<void> specialization - formatter improvements - Add incomplete std::set implementation - Lots of libc stdio progress - And more...

+++ boot/tui_efi.cc
@@ -84,7 +84,7 @@ private:
     T value;
 };
 
-static simple_atomic<int> systime_counter;
+static simple_atomic<uint64_t> systime_counter;
 
 static constexpr int keybuf_mask = 15;
 static int keybuf_queue[keybuf_mask + 1];
@@ -216,7 +216,7 @@ mouse_evt readmouse()
 }
 
 // Get ticks since midnight (54.9ms units)
-int systime()
+int64_t systime()
 {
     return systime_counter.get();
 }
